requirements.txt
anyio==3.7.0
bcrypt==4.0.1
certifi==2023.5.7
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
psycopg2==2.9.6
psycopg2-binary==2.9.6
pydantic==1.10.9
PyJWT==2.7.0
python-ulid==1.1.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
